content/canvas/test/test_text.font.html

<!DOCTYPE HTML>
<title>Canvas test: text.font</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

is(ctx.font, '10px sans-serif', "default font is not '10px sans-serif'");

ctx.save();
ctx.font = '20pt serif';
is(ctx.font, '20pt serif', 'font getter returns incorrect value');

ctx.restore();
is(ctx.font, '10px sans-serif', 'font not being stored in the context state');

if (!_deferred) SimpleTest.finish();
});
</script>
